added a signout controller
